refactor: Move NS install code into own module
Moves the code in question out from main into its own module
