+++ src-tauri/src/northstar/mod.rs
@@ -1,5 +1,6 @@
 //! This module deals with handling things around Northstar such as
 //! - getting version number
+pub mod install;
 
 use crate::{check_origin_running, constants::CORE_MODS, get_host_os, GameInstall, InstallType};
 use anyhow::anyhow;
